requirements.txt
antlr4-python3-runtime
blockdiag
colormap
coverage
dbfread
easydev
folium
jyquickhelper
keyring
Mako
networkx
openpyxl
pandas
git+https://github.com/pydata/pandas-datareader.git
pandocfilters
pymyinstall
pyquickhelper>=1.9
qgrid
quandl
scikit-learn
scipy
seaborn
sphinxjp.themes.basicstrap
wheel
yahoo-historical
